Running `oc` with Proc to avoid command expansion
This PR reimplements the mechanism to run `oc` command on a slave.

When using OpenShift Pipeline DSL, `oc` process is started with `DurableTask`,
which is designed to run a bash script or batch script on a slave.
This is not ideal because when the command arguments include special characters
(like whitespaces, `#`, `!`, `;`), it's hard to escape them correctly.

The new implementation uses standard Jenkins API to start a `oc` process
on a slave. `stdout` and `stderr` outputs are piped to Jenkins master
so that polling is not needed.