JKNS-458: Add Konflux-compatible Dockerfile
Adding a Dockerfile for testing the Jenkins client plugin that does not
rely on the OpenShift CI registry for base images. This will pay
dividends when we eventually build the Jenkins images and plugins on
Konflux, as this Dockerfile only uses productized images published on
registry.redhat.io.

This also adds a .dockerignore to ensure generated or test content is
not present when developers build this image themselves locally.

Signed-off-by: Adam Kaplan <adam.kaplan@redhat.com>